moved Customer + and Sale + to the components area, fixed add customer form
