+++ client/src/components/Header/index.js
@@ -22,23 +22,23 @@ function Header() {
                 </Link>
                 {/* <img src={profileImage} className="my-2" style={{ maxWidth: "8%", width: "40%" }} alt="Colossal Closer" /> */}
 
-                <a href="#" data-target="mobile-demo" className="sidenav-trigger"><i className="material-icons">menu</i></a>
+                <a href="#" data-target="mobile-demo" className="sidenav-trigger"><i className="material-icons"></i></a>
                 <ul id="nav-mobile" className="right ">
                     {loggedIn ? (
                         <>
                             <li><a href="" onClick={logout}>Logout</a></li>
-                            <li>
+                            {/* <li>
                             <Link to ="/AddCustomer"
                             href="/AddCustomer">Customer +
                             </Link>
                             
-                        </li>
-                        <li>
+                        </li> */}
+                        {/* <li>
                             <Link to ="/AddSale"
                             href="/AddSale">Sale +
                             </Link>
                             
-                        </li>
+                        </li> */}
                         </>
                     ):(
                          // <li><a href="sign.html">Login</a></li>
@@ -64,16 +64,16 @@ function Header() {
                             </Link>
                         </li>
                         <li className={`mx-2 ${(pageSelected === "customers") && 'active'}`}>
-                            <Link to="/customers" data-testid="customers" onClick={() => setPageSelected("customers")}>
+                            <Link to="/AddCustomer" data-testid="customers" onClick={() => setPageSelected("customers")}>
                                 Customers
                             </Link>
                         </li>
                         <li className={`mx-2 ${(pageSelected === "sales") && 'active'}`}>
-                            <Link to="/sales" data-testid="sales" onClick={() => setPageSelected("sales")}>
+                            <Link to="/AddSale" data-testid="sales" onClick={() => setPageSelected("sales")}>
                                 Sales
                             </Link>
                         </li>
-                        <li className="mx-2">
+                        {/* <li className="mx-2">
                             <Link to="/customer-graphs" data-testid="customerGraphs" onClick={() => setPageSelected("customerGraphs")}>
                                 Customers Graph
                             </Link>
@@ -82,7 +82,7 @@ function Header() {
                             <Link to="sales-graph" data-testid="salesGraph" onClick={() => setPageSelected("salesGraph")}>
                                 Sales Graph
                             </Link>
-                        </li>
+                        </li> */}
                     </ul>
                 </ul>
             </div>
