--- client/src/pages/AddCustomer/index.js
@@ -1,129 +1,0 @@
-import React, {useState} from 'react'
-import {useMutation} from '@apollo/react-hooks';
-import {ADD_CUSTOMER} from '../../utils/mutations';
-import Auth from '../../utils/auth';
-
-const AddCustomer = () => {
-    const [formState, setFormState] = useState({ businessName: '', contactName: '', phone: '', email: '', status: ''})
-
-    const [addCustomer, { error }] = useMutation(ADD_CUSTOMER);
-
-    const handleChange = (event) =>{
-        const {name,value} = event.target
-
-        setFormState({
-            ...formState,
-            [name]: value
-        });
-    };
-
-    const handleSignup = async (event) =>{
-        event.preventDefault();
-        console.log('button clicked', formState)
-        
-        // use try/catch instead of promises to handle errors
-        try{
-        //execute addUser mutation and pass in variable data from form
-        const { data } = await addCustomer({
-            variables: { ...formState }
-        });
-        Auth.login(data.addCustomer.token);
-        } catch (e){
-        console.error(e);
-        }
-     
-        setFormState({
-            bussinessName : '', 
-            contactName: '',
-            phone: '',
-            email: '',
-            status: ''
-            
-        })
-        
-    }
-    return (
-        <section id="login-page">
-            <div className="container center ">
-                
-                <div className="row" id="form-wrapper">
-
-                    <div className="col s12 m7 ">
-                        <div className="card" id="signup-card">
-
-                            <div className="card-content">
-                                <div className="row">
-                                    <form className="col s12" id="signup-form"  onSubmit={handleSignup}>
-                                        <div className="row">
-                                            <div className="input-field col s12">
-                                                <input 
-                                                id="bussinessName" 
-                                                type="text" 
-                                                name="bussinessName"
-                                                value= {formState.bussinessName}
-                                                onChange={handleChange}
-                                                />
-                                                <label htmlFor="bussinessName">Business Name</label>
-                                            </div>
-                                            <div className="input-field col s12">
-                                                <input 
-                                                id="contactName" 
-                                                type="text" 
-                                                name="contactName"
-                                                value= {formState.contactName}
-                                                onChange={handleChange}
-                                                />
-                                                <label htmlFor="contactName">Contact Name</label>
-                                            </div>
-                                        </div>
-                                        <div className="row">
-                                            <div className="input-field col s12">
-                                                <input
-                                                    id="phone" 
-                                                    type="text"
-                                                    name="phone"
-                                                    value= {formState.phone}
-                                                    onChange={handleChange}
-                                                />
-                                                <label htmlFor="phone">Phone #</label>
-                                            </div>
-                                            <div className="input-field col s12">
-                                                <input
-                                                    id="email" 
-                                                    type="email"
-                                                    name="email"
-                                                    value= {formState.email}
-                                                    onChange={handleChange}
-                                                />
-                                                <label htmlFor="email">Email</label>
-                                            </div>
-                                        </div>
-
-
-                                        {/* <div className="row">
-                                            <div className="input-field col s12">
-                                            
-                                                <input
-                                                    id="status" 
-                                                    type="status"
-                                                    name="status"
-                                                    value= {formState.status}
-                                                    onChange={handleChange}
-                                                />
-                                                <label htmlFor="email">Status</label>
-                                            </div>
-                                        </div> */}
-
-                                        <button className="blue lighten-3 waves-effect waves-light btn" type="submit" >Add Customer</button>
-                                    </form>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </section>
-    )
-}
-
-export default AddCustomer;
